pyproject.toml
[tool.poetry]
name = "global-search"
version = "0.1.0"
description = "Django based API application"
authors = ["Arun Kumar <arun.ku9555@gmail.com>"]
readme = "README.md"
packages = [{include = "global_search"}]

[tool.poetry.dependencies]
python = "^3.8"
django = "4.0"
django-environ = "0.8"
mysqlclient = "^2.2.0"
djangorestframework = "3.13.1"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"
pre-commit = "^3.3.3"


[tool.black]
# File filtering is taken care of in pre-commit.
line-length = 100
target-version = ["py38"]
exclude = '''
/(
    \.git
  | \.hg
  | \.mypy_cache
  | \.tox
  | \.venv
  | _build
  | buck-out
  | build
  | dist
  | migrations
)/
'''

[tool.isort]
# To know more about isort options
# https://pycqa.github.io/isort/docs/configuration/options.html
profile = "black"
line_length = 100
lines_between_sections = 1
multi_line_output = 3
include_trailing_comma = true
force_grid_wrap = 0
use_parentheses = true
src_paths = ["src"]
extend_skip = ["migrations"]

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
